# src/main/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: api-education-finder
  contact: { }
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "3.0"
servers:
  - url: http://localhost:56187
    description: Generated server url
...
paths:
  /{municipalityId}/statistics:
    get:
      tags:
        - Statistics
      description: |
        Find statistics based on the provided parameters.

        Filter parameters:
        - categories: Which categories to include in the statistics. Parameter filter strategy: IN-IGNORE-CASE. The parameter can be provided multiple times to specify multiple values. Example usage: ?categories=Teknik&categories=Bygg%20och%20anläggning
        - subcategories: Which subcategories to include in the statistics. Parameter filter strategy: IN-IGNORE-CASE. The parameter can be provided multiple times to specify multiple values. Example usage: ?subcategories=fysik&subcategories=matematik
        - levels: Which levels to include in the statistics. Parameter filter strategy: IN-IGNORE-CASE. The parameter can be provided multiple times to specify multiple values. Example usage: ?levels=gymnasial%20vuxenutbildning&levels=grundläggande%20vuxenutbildning
        - studyLocations: Which study locations to include in the statistics. Parameter filter strategy: IN-IGNORE-CASE. The parameter can be provided multiple times to specify multiple values. Example usage: ?studyLocations=Sundsvall&studyLocations=Östersund
        - scopes: Which scopes to include in the statistics. Parameter filter strategy: IN-IGNORE-CASE. The parameter can be provided multiple times to specify multiple values. Example usage: ?scopes=100&scopes=75
        - startDate: The start date of the statistics. 'startDate' is used in parallel with 'endDate' to form a period. Parameter filter strategy: WITHIN-OR-EQUAL. Example usage: ?startDate=2025-01-01
        - endDate: The end date of the statistics. 'endDate' is used in parallel with 'startDate' to form a period. Parameter filter strategy: WITHIN-OR-EQUAL. Example usage: ?endDate=2025-12-31

        Parameter filter strategies:
        - IN-IGNORE-CASE: Accepts a list of values separated by comma. The parameter value must have an exact match, case insensitive.
        - WITHIN-OR-EQUAL: The parameter value must be within the range of the searched attribute or equal to it. Example - startDate=2025-01-01, endDate=2025-12-31: The statistics will include all courses that have a start date between(inclusive) 2025-01-01 and 2025-12-31.

        Returned statistics:
        - OnGoingCourses: Number of courses which pass the filter parameters and have a start date before the given 'startDate' and after the given 'endDate'.
        - PlannedCourses: Number of courses which pass the filter parameters and have a start date after the given 'startDate' and before the given 'endDate'.
        - FinishedCourses: Number of courses which pass the filter parameters and have an end date after the given 'startDate' and before the given 'endDate'.
        - AvailableSeats: Number of available seats for all courses which pass the filter parameters and have a start date after the given 'startDate' and before the given 'endDate'.
        - TotalCapacity: Total capacity for all courses which pass the filter parameters and have a start date after the given 'startDate' and before the given 'endDate'.

        Important notes:
        - All date parameters are expected to be on the format: yyyy-MM-dd (example value: '2025-01-01')
      operationId: getStatistics
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: categories
          in: query
          description: List of categories
          required: false
          schema:
            type: string
            description: List of categories
            example: Medicin och vård
          example: Medicin och vård
        - name: subCategories
          in: query
          description: List of sub categories
          required: false
          schema:
            type: string
            description: List of sub categories
            example: PSYKOTERAPI
          example: PSYKOTERAPI
        - name: levels
          in: query
          description: List of levels
          required: false
          schema:
            type: string
            description: List of levels
            example: gymnasial vuxenutbildning
          example: gymnasial vuxenutbildning
        - name: studyLocations
          in: query
          description: List of study locations
          required: false
          schema:
            type: string
            description: List of study locations
            example: Sundsvall
          example: Sundsvall
        - name: scopes
          in: query
          description: List of scopes
          required: false
          schema:
            type: string
            description: List of scopes
            example: "25, 50, 75"
          example: "25, 50, 75"
        - name: startDate
          in: query
          description: Start date
          required: true
          schema:
            type: string
            description: Start date
            example: 2024-01-01
          example: 2024-01-01
        - name: endDate
          in: query
          description: End date
          required: true
          schema:
            type: string
            description: End date
            example: 2024-12-31
          example: 2024-12-31
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Statistics"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Statistics"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/statistics/filters/{filterAttribute}/values:
    get:
      tags:
        - Statistics
      summary: Find available filter values
      operationId: findFilterValues
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: filterAttribute
          in: path
          description: The attribute name to get available values from
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
            application/problem+json:
              schema:
                type: array
                items:
                  type: string
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/courses/filters/{filterAttribute}/values:
    get:
      tags:
        - Courses
      summary: Find available filter values
      description: Find available filter values (for use in the 'find'-resource)
      operationId: findFilterValues_1
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: filterAttribute
          in: path
          description: The attribute name to get available values from
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
            application/problem+json:
              schema:
                type: array
                items:
                  type: string
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        reasonPhrase:
          type: string
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    Statistics:
      type: object
      properties:
        onGoingCourses:
          type: integer
          description: Number of on-going courses
          format: int32
          readOnly: true
          example: 10
        plannedCourses:
          type: integer
          description: Number of planned courses
          format: int32
          readOnly: true
          example: 10
        finishedCourses:
          type: integer
          description: Number of finished courses
          format: int32
          readOnly: true
          example: 10
        availableSeats:
          type: integer
          description: Number of available seats
          format: int32
          readOnly: true
          example: 10
        totalCapacity:
          type: integer
          description: Total capacity
          format: int32
          readOnly: true
          example: 10
        studyLocations:
          type: array
          description: Study locations used for filtering
          readOnly: true
          items:
            type: string
            description: Study locations used for filtering
            readOnly: true
        scopes:
          type: array
          description: Scopes used for filtering
          readOnly: true
          items:
            type: integer
            description: Scopes used for filtering
            format: int32
            readOnly: true
        levels:
          type: array
          description: Levels used for filtering
          readOnly: true
          items:
            type: string
            description: Levels used for filtering
            readOnly: true
        categories:
          type: array
          description: Categories used for filtering
          readOnly: true
          items:
            type: string
            description: Categories used for filtering
            readOnly: true
        subCategories:
          type: array
          description: Category ids used for filtering
          readOnly: true
          items:
            type: string
            description: Category ids used for filtering
            readOnly: true
        startDate:
          type: string
          description: Start date used for filtering
          format: date
          readOnly: true
        endDate:
          type: string
          description: End date used for filtering
          format: date
          readOnly: true
      description: Statistics model
      readOnly: true
...
  securitySchemes: { }